requirements.txt
flask
pymongo
werkzeug
llama-cpp-python
huggingface_hub
